+++ src/tree/tree_node.rs
@@ -1,9 +1,6 @@
-#[cfg(any(feature = "use_serde", feature = "use_bincode", feature = "use_json", feature = "use_cbor", feature = "use_yaml", feature = "use_pickle", feature = "use_ron"))]
-use std::error::Error;
-
 use crate::merkle_bit::BinaryMerkleTreeResult;
 use crate::merkle_bit::NodeVariant;
-use crate::traits::{Exception, Node};
+use crate::traits::Node;
 use crate::tree::tree_branch::TreeBranch;
 use crate::tree::tree_leaf::TreeLeaf;
 use crate::tree::tree_data::TreeData;
@@ -30,14 +27,14 @@ use ron;
 #[cfg_attr(any(feature = "use_serde", feature = "use_bincode", feature = "use_json", feature = "use_cbor", feature = "use_yaml", feature = "use_pickle", feature = "use_ron"), derive(Serialize, Deserialize))]
 pub struct TreeNode {
     references: u64,
-    node: Option<NodeVariant<TreeBranch, TreeLeaf, TreeData>>,
+    node: NodeVariant<TreeBranch, TreeLeaf, TreeData>,
 }
 
 impl TreeNode {
-    fn new() -> Self {
+    fn new(node_variant: NodeVariant<TreeBranch, TreeLeaf, TreeData>) -> Self {
         Self {
             references: 0,
-            node: None,
+            node: node_variant,
         }
     }
 
@@ -49,27 +46,27 @@ impl TreeNode {
         self.references = references;
     }
     fn set_branch(&mut self, branch: TreeBranch) {
-        self.node = Some(NodeVariant::Branch(branch));
+        self.node = NodeVariant::Branch(branch);
     }
 
     fn set_leaf(&mut self, leaf: TreeLeaf) {
-        self.node = Some(NodeVariant::Leaf(leaf));
+        self.node = NodeVariant::Leaf(leaf);
     }
     fn set_data(&mut self, data: TreeData) {
-        self.node = Some(NodeVariant::Data(data));
+        self.node = NodeVariant::Data(data);
     }
 }
 
 #[cfg(feature = "use_bincode")]
 impl Encode for TreeNode {
-    fn encode(&self) -> Result<Vec<u8>, Box<Error>> {
+    fn encode(&self) -> BinaryMerkleTreeResult<Vec<u8>> {
         Ok(serialize(self)?)
     }
 }
 
 #[cfg(feature = "use_json")]
 impl Encode for TreeNode {
-    fn encode(&self) -> Result<Vec<u8>, Box<Error>> {
+    fn encode(&self) -> BinaryMerkleTreeResult<Vec<u8>> {
         let encoded = serde_json::to_string(&self)?;
         Ok(encoded.as_bytes().to_vec())
     }
@@ -77,42 +74,42 @@ impl Encode for TreeNode {
 
 #[cfg(feature = "use_cbor")]
 impl Encode for TreeNode {
-    fn encode(&self) -> Result<Vec<u8>, Box<Error>> {
+    fn encode(&self) -> BinaryMerkleTreeResult<Vec<u8>> {
         Ok(serde_cbor::to_vec(&self)?)
     }
 }
 
 #[cfg(feature = "use_yaml")]
 impl Encode for TreeNode {
-    fn encode(&self) -> Result<Vec<u8>, Box<Error>> {
+    fn encode(&self) -> BinaryMerkleTreeResult<Vec<u8>> {
         Ok(serde_yaml::to_vec(&self)?)
     }
 }
 
 #[cfg(feature = "use_pickle")]
 impl Encode for TreeNode {
-    fn encode(&self) -> Result<Vec<u8>, Box<Error>> {
+    fn encode(&self) -> BinaryMerkleTreeResult<Vec<u8>> {
         Ok(serde_pickle::to_vec(&self, true)?)
     }
 }
 
 #[cfg(feature = "use_ron")]
 impl Encode for TreeNode {
-    fn encode(&self) -> Result<Vec<u8>, Box<Error>> {
+    fn encode(&self) -> BinaryMerkleTreeResult<Vec<u8>> {
         Ok(ron::ser::to_string(&self)?.as_bytes().to_vec())
     }
 }
 
 #[cfg(feature = "use_bincode")]
 impl Decode for TreeNode {
-    fn decode(buffer: &[u8]) -> Result<Self, Box<Error>> {
+    fn decode(buffer: &[u8]) -> BinaryMerkleTreeResult<Self> {
         Ok(deserialize(buffer)?)
     }
 }
 
 #[cfg(feature = "use_json")]
 impl Decode for TreeNode {
-    fn decode(buffer: &[u8]) -> Result<Self, Box<Error>> {
+    fn decode(buffer: &[u8]) -> BinaryMerkleTreeResult<Self> {
         let decoded_string = String::from_utf8(buffer.to_vec())?;
         let decoded = serde_json::from_str(&decoded_string)?;
         Ok(decoded)
@@ -121,43 +118,39 @@ impl Decode for TreeNode {
 
 #[cfg(feature = "use_cbor")]
 impl Decode for TreeNode {
-    fn decode(buffer: &[u8]) -> Result<Self, Box<Error>> {
+    fn decode(buffer: &[u8]) -> BinaryMerkleTreeResult<Self> {
         Ok(serde_cbor::from_slice(buffer)?)
     }
 }
 
 #[cfg(feature = "use_yaml")]
 impl Decode for TreeNode {
-    fn decode(buffer: &[u8]) -> Result<Self, Box<Error>> {
+    fn decode(buffer: &[u8]) -> BinaryMerkleTreeResult<Self> {
         Ok(serde_yaml::from_slice(buffer)?)
     }
 }
 
 #[cfg(feature = "use_pickle")]
 impl Decode for TreeNode {
-    fn decode(buffer: &[u8]) -> Result<Self, Box<Error>> {
+    fn decode(buffer: &[u8]) -> BinaryMerkleTreeResult<Self> {
         Ok(serde_pickle::from_slice(buffer)?)
     }
 }
 
 #[cfg(feature = "use_ron")]
 impl Decode for TreeNode {
-    fn decode(buffer: &[u8]) -> Result<Self, Box<Error>> {
+    fn decode(buffer: &[u8]) -> BinaryMerkleTreeResult<Self> {
         Ok(ron::de::from_bytes(buffer)?)
     }
 }
 
 impl<ValueType> Node<TreeBranch, TreeLeaf, TreeData, ValueType> for TreeNode
     where ValueType: Encode + Decode {
-    fn new() -> Self { Self::new() }
+    fn new(node_variant: NodeVariant<TreeBranch, TreeLeaf, TreeData>) -> Self { Self::new(node_variant) }
 
     fn get_references(&self) -> u64 { Self::get_references(&self) }
-    fn get_variant(self) -> BinaryMerkleTreeResult<NodeVariant<TreeBranch, TreeLeaf, TreeData>> {
-        if let Some(variant) = self.node {
-            Ok(variant)
-        } else {
-            return Err(Box::new(Exception::new("Failed to distinguish node type")))
-        }
+    fn get_variant(self) -> NodeVariant<TreeBranch, TreeLeaf, TreeData> {
+        self.node
     }
 
     fn set_references(&mut self, references: u64) { Self::set_references(self, references) }